# apiserver/model_booking.go
/*
 * Booking App
 *
 * API to access and configure the Booking App
 *
 * API version: 1.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package apiserver

import (
	"time"
)

type Booking struct {

	// The booking ID.
	Id string `json:"id,omitempty"`

	// The start datetime of the booking in ISO 8601 format.
	Start time.Time `json:"start,omitempty"`

	// The end datetime of the booking in ISO 8601 format.
	End time.Time `json:"end,omitempty"`

	// The ID (email) of the organizer.
	OrganizerID string `json:"organizerID,omitempty"`

	// The name of the organizer.
	OrganizerName string `json:"organizerName,omitempty"`
}

// AssertBookingRequired checks if the required fields are not zero-ed
func AssertBookingRequired(obj Booking) error {
	return nil
}

// AssertBookingConstraints checks if the values respects the defined constraints
func AssertBookingConstraints(obj Booking) error {
	return nil
}